amended generateReportController.ts amended generateReportController.test.ts amended controllers/index.ts amended routes/index.ts deleted providerReportController.ts deleted providerReportController.test.ts
