+++ server/controllers/generateReportController.ts
@@ -12,19 +12,23 @@ const VIEW_PATH = 'pages/generateReport'
 export default class GenerateReportController {
   constructor(private readonly generateReportService: GenerateReportService) {}
 
-  show(): RequestHandler {
+  show(isProviderReport = false): RequestHandler {
     return async (req: Request, res: Response): Promise<void> => {
       const backUrl = manageBackLink(CURRENT_URL)
       res.render(VIEW_PATH, {
         backUrl,
-        isProviderReport: false,
+        isProviderReport,
+        formValues: {},
+        errors: {},
       })
     }
   }
 
   submit(): RequestHandler {
     return async (req: Request, res: Response): Promise<void> => {
-      const reportParams: Record<string, string> = {
+      const isProviderReport = req.path === '/provider-report'
+
+      const allParams: Record<string, string> = {
         crmType: req.body.crmType as string,
         decisionFromDate: req.body.decisionFromDate as string,
         decisionToDate: req.body.decisionToDate as string,
@@ -34,23 +38,30 @@ export default class GenerateReportController {
         createdToDate: req.body.createdToDate as string,
         lastSubmittedFromDate: req.body.lastSubmittedFromDate as string,
         lastSubmittedToDate: req.body.lastSubmittedToDate as string,
+        providerAccount: req.body.providerAccount as string,
       }
-      const validationErrors = validateReportParams(reportParams)
 
+      const reportParams = this.buildReportParams(allParams, isProviderReport)
+
+      const validationErrors = validateReportParams(reportParams, isProviderReport)
       if (validationErrors) {
         // render with validation errors
         res.render(VIEW_PATH, {
-          results: [],
+          formValues: allParams,
           errors: validationErrors,
-          formValues: reportParams,
           backUrl: manageBackLink(CURRENT_URL),
-          isProviderReport: false,
+          isProviderReport,
         })
-      } else {
-        // perform generate report
-        const crmReportRequest = this.buildReportRequest(reportParams, getProfileAcceptedTypes(res))
-        const crmReportResponse = await this.generateReportService.getCrmReport(crmReportRequest)
+        return
+      }
 
+      // perform generate report
+      const crmReportRequest = this.buildReportRequest(reportParams, getProfileAcceptedTypes(res), isProviderReport)
+
+      try {
+        const crmReportResponse = isProviderReport
+          ? await this.generateReportService.getProviderCrmReport(crmReportRequest)
+          : await this.generateReportService.getCrmReport(crmReportRequest)
         // check for any errors in the report response
         if (crmReportResponse.errorMessage) {
           // render with errors for report API error
@@ -60,18 +71,59 @@ export default class GenerateReportController {
             errors,
             formValues: reportParams,
             backUrl: manageBackLink(CURRENT_URL),
-            isProviderReport: false,
+            isProviderReport,
           })
-        } else {
-          res.setHeader('Content-Type', 'text/csv')
-          res.setHeader('Content-Disposition', `attachment; filename=${this.getReportFilename(reportParams.crmType)}`)
-          res.send(crmReportResponse.text)
+          return
         }
+
+        // Send the report as a CSV
+        res.setHeader('Content-Type', 'text/csv')
+        res.setHeader(
+          'Content-Disposition',
+          `attachment; filename=${this.getReportFilename(reportParams.crmType, isProviderReport)}`,
+        )
+        res.send(crmReportResponse.text)
+      } catch (error) {
+        // Handle unexpected errors
+        res.render(VIEW_PATH, {
+          errors: { list: [{ href: '#', text: 'An unexpected error occurred. Please try again later.' }] },
+          formValues: allParams,
+          backUrl: manageBackLink(CURRENT_URL),
+          isProviderReport,
+        })
       }
     }
   }
 
-  private buildReportRequest(reportParams: Record<string, string>, profileAcceptedTypes: string): CrmReportRequest {
+  private buildReportParams(allParams: Record<string, string>, isProviderReport: boolean): Record<string, string> {
+    const reportParams: Record<string, string> = { crmType: allParams.crmType }
+
+    // Include decision dates for all CRM types
+    reportParams.decisionFromDate = allParams.decisionFromDate
+    reportParams.decisionToDate = allParams.decisionToDate
+
+    if (isProviderReport && allParams.crmType === 'crm4') {
+      // Only include providerAccount for provider reports with CRM4
+      reportParams.providerAccount = allParams.providerAccount
+    }
+
+    if (allParams.crmType === 'crm14') {
+      reportParams.submittedFromDate = allParams.submittedFromDate
+      reportParams.submittedToDate = allParams.submittedToDate
+      reportParams.createdFromDate = allParams.createdFromDate
+      reportParams.createdToDate = allParams.createdToDate
+      reportParams.lastSubmittedFromDate = allParams.lastSubmittedFromDate
+      reportParams.lastSubmittedToDate = allParams.lastSubmittedToDate
+    }
+
+    return reportParams
+  }
+
+  private buildReportRequest(
+    reportParams: Record<string, string>,
+    profileAcceptedTypes: string,
+    isProviderReport: boolean,
+  ): CrmReportRequest {
     return {
       crmType: reportParams.crmType,
       decisionFromDate: reportParams.decisionFromDate,
@@ -82,11 +134,12 @@ export default class GenerateReportController {
       createdToDate: reportParams.createdToDate,
       lastSubmittedFromDate: reportParams.lastSubmittedFromDate,
       lastSubmittedToDate: reportParams.lastSubmittedToDate,
-      profileAcceptedTypes,
+      providerAccount: isProviderReport ? reportParams.providerAccount : undefined, // Include only for provider reports
+      profileAcceptedTypes: isProviderReport ? '' : profileAcceptedTypes, // Set profileAcceptedTypes for non-provider reports
     }
   }
 
-  private getReportFilename(crmType: string): string {
-    return `${crmType}Report.csv`
+  private getReportFilename(crmType: string, isProviderReport: boolean): string {
+    return `${crmType}${isProviderReport ? '-Provider' : ''}Report.csv`
   }
 }
